Manage and find obsolete file types in the system
Allows "obsolete" file types (by extension) to be managed in the system. Once they are input a listing of all the obsolete types can be seen.

From this point it is a policy/procedure to convert them (either replace them in the system with a converted version or define a conversion to a known type and keey the conversion in the exports.).

Keep in mind that when "overwriting" a file with a newer version the original "obsolete" version will still exist in revision control. It is recommended that you download, verify via check sum, and then upload the newer version of the file as the archival copy. If you ever need to go back to the original (now obsolete version) use revision control.