Add formCreator permissions to the MFCS users.
The acl.php has been modified to use the locally defined permissions instead of the ones from the Active Directory.

We now load all the MFCS classes and the MFCS singleton before the inclusion of the acl.php files (so that we can use the mfcsPerms class in the acl).