Add Public Export and OAI form metadata options
Add two new options to the form to designate as exporting to public, and designate the form as OAI. These are required to properly export metadata to the OAI server. Otherwise we do not know if 1) a form is ready for release to the public, and 2) if a form is supposed to be indexed in the OAI server for ppublic harvisting.

Having both options, plus individual exporting options n each record, it allows for fine tuning the automatic OAI harvisting.

The exporting class should be updated to make use of export to public. 

This feature differs from "Production Ready" in that production ready allows staff to enter data into the form. Allow release to public is when the form is ready for exporting as public consumption.