Remove spaces around the joining of matadata standards
Spaces were incorrectly used in the concatination of the metadata standards information. This removes them on creation. 

It does not address the issue that there is already data saved incorrectly in production. WVU is not currently using the metadata information in production. 

To mitigate this issue the best course of action will be to open all the forms that have metadata standards defined and save them. This should save the data correctly and not require aditional modifications.