Ad OAI Release to individual fields.
This is required because some fields may be tagged for public release, to a local repository, but should not be included in an OAI release. this could be the case, especially, if there are multiple digital object fields. The public release can be utilized to export all digital objects, but then be hidden from OAI and a single URL can be provided for the item.