Add OAI support & Bug fixes
* Add Public Export and OAI form metadata options

Add two new options to the form to designate as exporting to public, and designate the form as OAI. These are required to properly export metadata to the OAI server. Otherwise we do not know if 1) a form is ready for release to the public, and 2) if a form is supposed to be indexed in the OAI server for ppublic harvisting.

Having both options, plus individual exporting options n each record, it allows for fine tuning the automatic OAI harvisting.

The exporting class should be updated to make use of export to public. 

This feature differs from "Production Ready" in that production ready allows staff to enter data into the form. Allow release to public is when the form is ready for exporting as public consumption.

* nagios check

* Change the temporary location where the download all is built

* Add objects::getByIDNO()

Retrieve an object with its IDNO, instead of the system ObjectID

* API for adding a URL to an Object

This is provided so that import scripts for other systems ca update MFCS with regards to where the object is viewable by the public.

* Switched all methods in Objects class to use standard $engine call

* Updated securityConcerns with clarifying information.

* Changed objects::getObjects() to honor ignoreCache

* Add method to count all objects in the system.

* Add column to objects table for publicRelease of objects

* CLI Script to convert from previous way of defining publicRelease object

* Add Release to Public to Object forms

1. Add ability to add a public release select box to object forms.
1. Add ability to set the default value of yes/no to public release select
1. add public release to the object create/update forms

Note: when this gets merged into master someone will have to update all of the forms and the export forms. Any form with "publicRelease" will need that removed from the form. 

Any exports that use public release will need to be updated.

* Add OAI Release to individual fields. 

This is required because some fields may be tagged for public release, to a local repository, but should not be included in an OAI release. this could be the case, especially, if there are multiple digital object fields. The public release can be utilized to export all digital objects, but then be hidden from OAI and a single URL can be provided for the item.

* Display Public URLs on the object page, in a new tab.

* Starting unit testing framework. Update for public release.

1. Created new class string utils
1. Added todos for removing, refactoring

* readme file for put url API.

* Only display the Public Release select on object forms

* Add API Endpoint for getting a URL for an object

* Increase the length of the objectUrl tables

* Fix for drive usage on dashboard.

* Projects api endpoint (#264)

* Add endpoint to pull Object IDNOs that belong to a project

* Projects api endpoint (#266)

* Add endpoint to pull Object IDNOs that belong to a project

* Path correction

* Correct $engine variable

* Check for existence of project

* Changing JavaScript to reveal fields. (#265)

* Fix incorrect number of fields on form insert

* Display selected option text instead of ID number, for select fields in metadata forms

* Bug FIx: Correct issue where adding an element to a metadata form would remove it from production